requirements.txt
# requirements.txt
# usage: pip install -r requirements.txt

Django==5.1
djangorestframework==3.15.1
django-crispy-forms==2.3
crispy-bootstrap4==2024.1
django-select2==8.2.1
jira==3.8.0
requests=2.32.3

# PostgreSQL: psycopg2 or psycopg2-binary for Windows
# LDAP authentication: django-python3-ldap

# CI
black==24.8.0
flake8==7.1.1
isort==5.13.2
pre-commit==3.8.0
